requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-dotenv==1.0.0
Pillow==10.1.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-filter==23.3
pytest==7.4.3
pytest-django==4.7.0
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4